fix: update extension only behaviour and closeicon size
Closes #338
