+++ packages/screens/src/send/NFTListScreen.query.tsx
@@ -492,6 +492,7 @@ export function NFTListScreen(): React.ReactElement {
           })()}
           isEditing={false}
           onQuantityChange={handleQuantityChange}
+          isExtension={isExtension}
         />
       </YStack>
     </BackgroundWrapper>
