+++ packages/ui/src/components/NFTSelectionBar.tsx
@@ -19,6 +19,7 @@ export interface NFTSelectionBarProps {
   // Text props for localization
   selectedCountText?: string;
   confirmText?: string;
+  isExtension?: boolean;
 }
 
 export function NFTSelectionBar({
@@ -29,6 +30,7 @@ export function NFTSelectionBar({
   onQuantityChange,
   selectedCountText,
   confirmText,
+  isExtension,
 }: NFTSelectionBarProps) {
   const theme = useTheme();
   const isCurrentlyDarkMode = isDarkMode(theme);
@@ -136,7 +138,7 @@ export function NFTSelectionBar({
 
   return (
     <YStack
-      pos="absolute"
+      pos={isExtension ? 'fixed' : 'absolute'}
       b="$0"
       l="$0"
       r="$0"
